requirements.txt
Flask==1.0.2
python-dotenv==0.8.2
PyJWT==1.6.4